[GFC][Cleanup] Move type aliases to dedicated header. https://bugs.webkit.org/show_bug.cgi?id=299615 rdar://problem/161422019
Reviewed by Brent Fulgham.

We have various type alises that we use throughout different files in
GFC. Sometimes this results in a bunch of redundant declarations
when the type alias or some other forward declaration is not included.
If we move these type alises along with the forward declarations needed
for them into a dedicated header file this significantly reduces the
when reading the rest of the code.

Along with the change above there are also a couple of other related
changes:
1.) Move the UsedTrackSizes and UnsizedTrack struct definitions from
the headers to the their respective sources since they are currently
not used outside of thier translation units.

2.) Move TrackSizingFunctions to its own header.

Canonical link: https://commits.webkit.org/300604@main
